+++ src/meta/src/manager/env.rs
@@ -16,14 +16,11 @@ use std::ops::Deref;
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use anyhow::Context;
-use notify::Watcher;
 use risingwave_common::config::{
     CompactionConfig, DefaultParallelism, MetaBackend, ObjectStoreConfig,
 };
 use risingwave_common::session_config::SessionConfig;
 use risingwave_common::system_param::reader::SystemParamsReader;
-use risingwave_common::system_param::LICENSE_KEY_KEY;
 use risingwave_meta_model_migration::{MigrationStatus, Migrator, MigratorTrait};
 use risingwave_meta_model_v2::prelude::Cluster;
 use risingwave_pb::meta::SystemParams;
@@ -31,8 +28,6 @@ use risingwave_rpc_client::{
     FrontendClientPool, FrontendClientPoolRef, StreamClientPool, StreamClientPoolRef,
 };
 use sea_orm::EntityTrait;
-use thiserror_ext::AsReport;
-use tokio::sync::watch;
 
 use super::{
     SessionParamsManager, SessionParamsManagerRef, SystemParamsManager, SystemParamsManagerRef,
@@ -634,79 +629,5 @@ impl MetaSrvEnv {
         )
         .await
         .unwrap()
-    }
-}
-
-impl MetaSrvEnv {
-    /// Spawn background tasks to watch the license key file and update the system parameter,
-    /// if configured.
-    pub fn may_start_watch_license_key_file(&self) -> MetaResult<()> {
-        let Some(path) = self.opts.license_key_path.as_ref() else {
-            return Ok(());
-        };
-
-        let (changed_tx, mut changed_rx) = watch::channel(());
-
-        let mut watcher =
-            notify::recommended_watcher(move |event: Result<notify::Event, notify::Error>| {
-                if let Err(e) = event {
-                    tracing::warn!(
-                        error = %e.as_report(),
-                        "error occurred while watching license key file"
-                    );
-                    return;
-                }
-                let _ = changed_tx.send(());
-            })
-            .context("failed to create license key file watcher")?;
-
-        // This will spawn a new thread to watch the file, so no need to be concerned about blocking.
-        watcher
-            .watch(path, notify::RecursiveMode::NonRecursive)
-            .context("failed to watch license key file")?;
-
-        let updater = {
-            let mgr = self.system_param_manager_impl.clone();
-            let path = path.to_path_buf();
-            async move {
-                // Let the watcher live until the end of the updater to prevent dropping (then stopping).
-                let _watcher = watcher;
-
-                while changed_rx.changed().await.is_ok() {
-                    let content = match tokio::fs::read_to_string(&path).await {
-                        Ok(v) => v,
-                        Err(e) => {
-                            tracing::warn!(
-                                path = %path.display(),
-                                error = %e.as_report(),
-                                "failed to read license key file"
-                            );
-                            continue;
-                        }
-                    };
-                    let content = content.trim().to_owned();
-                    let value = (!content.is_empty()).then_some(content);
-
-                    let result = match &mgr {
-                        SystemParamsManagerImpl::Kv(mgr) => {
-                            mgr.set_param(LICENSE_KEY_KEY, value).await
-                        }
-                        SystemParamsManagerImpl::Sql(mgr) => {
-                            mgr.set_param(LICENSE_KEY_KEY, value).await
-                        }
-                    };
-
-                    if let Err(e) = result {
-                        tracing::error!(
-                            error = %e.as_report(),
-                            "failed to set license key from file"
-                        );
-                    }
-                }
-            }
-        };
-        let _handle = tokio::spawn(updater);
-
-        Ok(())
     }
 }
